Moved handler classes to package internal.handlers
